@@ -1,78 +1,71 @@
 {
-  "name": "pumba-uploader",
-  "version": "0.0.1",
-  "description": "pumba-uploader",
-  "keywords": [
-    "loopback-application",
-    "loopback"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "engines": {
-    "node": ">=10.16"
-  },
-  "scripts": {
-    "build": "lb-tsc",
-    "build:watch": "lb-tsc --watch",
-    "lint": "npm run eslint && npm run prettier:check",
-    "lint:fix": "npm run eslint:fix && npm run prettier:fix",
-    "prettier:cli": "lb-prettier \"**/*.ts\" \"**/*.js\"",
-    "prettier:check": "npm run prettier:cli -- -l",
-    "prettier:fix": "npm run prettier:cli -- --write",
-    "eslint": "lb-eslint --report-unused-disable-directives .",
-    "eslint:fix": "npm run eslint -- --fix",
-    "pretest": "npm run rebuild",
-    "test": "lb-mocha --allow-console-logs \"dist/__tests__\"",
-    "posttest": "npm run lint",
-    "test:dev": "lb-mocha --allow-console-logs dist/__tests__/**/*.js && npm run posttest",
-    "docker:build": "docker build -t pumba-uploader .",
-    "docker:run": "docker run -p 3000:3000 -d pumba-uploader",
-    "premigrate": "npm run build",
-    "migrate": "node ./dist/migrate",
-    "preopenapi-spec": "npm run build",
-    "openapi-spec": "node ./dist/openapi-spec",
-    "rebuild": "npm run clean && npm run build",
-    "prestart": "npm run rebuild",
-    "start": "node -r source-map-support/register .",
-    "clean": "lb-clean dist *.tsbuildinfo .eslintcache"
-  },
-  "repository": {
-    "type": "git",
-    "url": ""
-  },
-  "author": "itamar-marom <marom.itamar@gmail.com>",
-  "license": "",
-  "files": [
-    "README.md",
-    "dist",
-    "src",
-    "!*/__tests__"
-  ],
-  "dependencies": {
-    "@loopback/boot": "^3.1.2",
-    "@loopback/core": "^2.13.1",
-    "@loopback/repository": "^3.3.0",
-    "@loopback/rest": "^9.1.1",
-    "@loopback/rest-explorer": "^3.0.5",
-    "@loopback/service-proxy": "^3.0.5",
-    "axios": "^0.21.1",
-    "cookie-signature": "^1.0.6",
-    "file-type": "^16.5.3",
-    "form-parser": "^0.1.1",
-    "got": "^12.0.1",
-    "http": "0.0.1-security",
-    "loopback-connector-rest": "^4.0.1",
-    "request": "^2.88.2",
-    "tslib": "^2.0.0"
-  },
-  "devDependencies": {
-    "@loopback/build": "^6.2.8",
-    "@loopback/eslint-config": "^10.0.4",
-    "@loopback/testlab": "^3.2.10",
-    "@types/node": "^10.17.48",
-    "@types/request": "^2.48.5",
-    "eslint": "^7.15.0",
-    "source-map-support": "^0.5.19",
-    "typescript": "~4.1.2"
-  }
+	"name": "pumba-uploader",
+	"version": "0.0.1",
+	"description": "pumba-uploader",
+	"keywords": [ "loopback-application", "loopback" ],
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"engines": {
+		"node": ">=10.16"
+	},
+	"scripts": {
+		"build": "lb-tsc",
+		"build:watch": "lb-tsc --watch",
+		"lint": "npm run eslint && npm run prettier:check",
+		"lint:fix": "npm run eslint:fix && npm run prettier:fix",
+		"prettier:cli": "lb-prettier \"**/*.ts\" \"**/*.js\"",
+		"prettier:check": "npm run prettier:cli -- -l",
+		"prettier:fix": "npm run prettier:cli -- --write",
+		"eslint": "lb-eslint --report-unused-disable-directives .",
+		"eslint:fix": "npm run eslint -- --fix",
+		"pretest": "npm run rebuild",
+		"test": "lb-mocha --allow-console-logs \"dist/__tests__\"",
+		"posttest": "npm run lint",
+		"test:dev": "lb-mocha --allow-console-logs dist/__tests__/**/*.js && npm run posttest",
+		"docker:build": "docker build -t pumba-uploader .",
+		"docker:run": "docker run -p 3000:3000 -d pumba-uploader",
+		"premigrate": "npm run build",
+		"migrate": "node ./dist/migrate",
+		"preopenapi-spec": "npm run build",
+		"openapi-spec": "node ./dist/openapi-spec",
+		"rebuild": "npm run clean && npm run build",
+		"prestart": "npm run rebuild",
+		"start": "node -r source-map-support/register .",
+		"clean": "lb-clean dist *.tsbuildinfo .eslintcache"
+	},
+	"repository": {
+		"type": "git",
+		"url": ""
+	},
+	"author": "itamar-marom <marom.itamar@gmail.com>",
+	"license": "",
+	"files": [ "README.md", "dist", "src", "!*/__tests__" ],
+	"dependencies": {
+		"@loopback/boot": "^3.1.2",
+		"@loopback/core": "^2.13.1",
+		"@loopback/repository": "^3.3.0",
+		"@loopback/rest": "^9.1.1",
+		"@loopback/rest-explorer": "^3.0.5",
+		"@loopback/service-proxy": "^3.0.5",
+		"axios": "^0.21.1",
+		"cookie-signature": "^1.0.6",
+		"file-type": "^16.5.3",
+		"form-parser": "^0.1.1",
+		"got": "^12.0.1",
+		"http": "0.0.1-security",
+		"loopback-connector-rest": "^4.0.1",
+		"request": "^2.88.2",
+		"tslib": "^2.0.0",
+		"typescript": "4.5.5"
+	},
+	"devDependencies": {
+		"@loopback/build": "^6.2.8",
+		"@loopback/eslint-config": "^10.0.4",
+		"@loopback/testlab": "^3.2.10",
+		"@types/node": "^10.17.48",
+		"@types/request": "^2.48.5",
+		"eslint": "^7.15.0",
+		"source-map-support": "^0.5.19",
+		"typescript": "4.5.5"
+	}
 }
