@@ -1,67 +1,67 @@
-// import { inject, lifeCycleObserver, LifeCycleObserver } from '@loopback/core';
-// import { juggler } from '@loopback/repository';
+import { inject, lifeCycleObserver, LifeCycleObserver } from '@loopback/core';
+import { juggler } from '@loopback/repository';
 
-// // TODO change these too
-// const NEXUS_URL = process.env.NEXUS_BASE_URL;
-// const AUTH_TOKEN = process.env.NEXUS_UPLOAD_AUTH_TOKEN;
-// const UPLOAD_URL = `/service/rest/v1/components?repository=`;
+// TODO change these too
+const NEXUS_URL = process.env.NEXUS_BASE_URL;
+const AUTH_TOKEN = process.env.NEXUS_UPLOAD_AUTH_TOKEN;
+// const AUTH_TOKEN = 'Basic YWRtaW46ZGV2b3BzNEVWRVI=';
 
-// if (NEXUS_URL == undefined || AUTH_TOKEN == undefined) {
-// 	throw new Error('base variables are undefined somehow -- nexusds');
-// }
+const UPLOAD_URL = `/service/rest/v1/components?repository=`;
 
-// const config = {
-// 	name: 'nexusds',
-// 	connector: 'rest',
-// 	baseURL: NEXUS_URL,
-// 	crud: false,
-// 	options: {
-// 		headers: {
-// 			accept: 'application/json'
-// 			// 'content-type': 'application/json'
-// 		}
-// 	},
-// 	// TODO -- change the function here for file upload
-// 	operations: [
-// 		{
-// 			template: {
-// 				method: 'POST',
-// 				url: NEXUS_URL + UPLOAD_URL + '{repository}',
-// 				headers: {
-// 					Authorization: AUTH_TOKEN,
-// 					accept: 'application/json',
-// 					'Content-Type': 'multipart/form-data',
-// 					'Content-Disposition': 'attachment; filename: {filename}'
-// 				},
-// 				body: {
-// 					'npm.asset': {
-// 						value: '{payload}',
-// 						type: 'type=application/x-compressed',
-// 						options: { filename: '{filename}', contentType: null }
-// 					}
-// 				},
-// 				fullResponse: true
-// 			},
-// 			functions: {
-// 				uploadFile: [ 'repository', 'payload', 'mimetype', 'filename' ]
-// 			}
-// 		}
-// 	]
-// };
+if (NEXUS_URL == undefined || AUTH_TOKEN == undefined) {
+	throw new Error('base variables are undefined somehow -- nexusds');
+}
 
-// // Observe application's life cycle to disconnect the datasource when
-// // application is stopped. This allows the application to be shut down
-// // gracefully. The `stop()` method is inherited from `juggler.DataSource`.
-// // Learn more at https://loopback.io/doc/en/lb4/Life-cycle.html
-// @lifeCycleObserver('datasource')
-// export class NexusdsDataSource extends juggler.DataSource implements LifeCycleObserver {
-// 	static dataSourceName = 'nexusds';
-// 	static readonly defaultConfig = config;
+const config = {
+	name: 'nexusds',
+	connector: 'rest',
+	baseURL: NEXUS_URL,
+	crud: false,
+	options: {
+		headers: {
+			accept: 'application/json'
+		}
+	},
+	operations: [
+		{
+			template: {
+				method: 'POST',
+				url: NEXUS_URL + UPLOAD_URL + '{repository}',
+				headers: {
+					Authorization: AUTH_TOKEN,
+					accept: 'application/json',
+					'Content-Type': 'multipart/form-data',
+					'Content-Disposition': 'attachment; filename={filename}'
+				},
+				formData: {
+					'npm.asset': {
+						value: '{payload}',
+						type: 'type=application/x-compressed',
+						options: { filename: '{filename}', contentType: null }
+					}
+				},
+				fullResponse: true
+			},
+			functions: {
+				uploadFile: [ 'repository', 'payload', 'filename' ]
+			}
+		}
+	]
+};
 
-// 	constructor(
-// 		@inject('datasources.config.nexusds', { optional: true })
-// 		dsConfig: object = config
-// 	) {
-// 		super(dsConfig);
-// 	}
-// }
+// Observe application's life cycle to disconnect the datasource when
+// application is stopped. This allows the application to be shut down
+// gracefully. The `stop()` method is inherited from `juggler.DataSource`.
+// Learn more at https://loopback.io/doc/en/lb4/Life-cycle.html
+@lifeCycleObserver('datasource')
+export class NexusdsDataSource extends juggler.DataSource implements LifeCycleObserver {
+	static dataSourceName = 'nexusds';
+	static readonly defaultConfig = config;
+
+	constructor(
+		@inject('datasources.config.nexusds', { optional: true })
+		dsConfig: object = config
+	) {
+		super(dsConfig);
+	}
+}
